+++ Diary/Model/Diary.swift
@@ -12,20 +12,3 @@ struct Diary: Decodable {
   let body: String
   let createdAt: TimeInterval
 }
-
-struct DiaryData {
-  static func saveJsonData() -> [Diary]? {
-    guard let path = Bundle.main.path(forResource: "sample", ofType: "json") else {
-      return nil
-    }
-    
-    guard let jsonData = try? String(contentsOfFile: path).data(using: .utf8) else {
-      return nil
-    }
-    
-    let diaryData = try? Json.decoder.decode([Diary].self, from: jsonData)
-    
-    return diaryData
-    
-  }
-}
